requirements.txt
asgiref==3.3.4
Django==3.1.4
djangorestframework==3.12.2
djangorestframework-simplejwt==4.4.0
pkg-resources==0.0.0
PyJWT==1.7.1
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
